refactor : Separate into a list component containing cafe information
